6241314 ctfconvert should understand SPARC real types in DWARF 6313293 ctfstabs is confused by gcc's output 6268862 add more gcc support routines to a few places 6272005 gcc and cmd/mdb don't get along 6273966 gcc and common/dis don't get along
--HG--
rename : usr/src/cmd/mdb/intel/ia32/libstand/muldiv.s => deleted_files/usr/src/cmd/mdb/intel/ia32/libstand/muldiv.s
rename : usr/src/cmd/mdb/intel/ia32/libstand/sys/archsalib.h => deleted_files/usr/src/cmd/mdb/intel/ia32/libstand/sys/archsalib.h

+++ usr/src/cmd/mdb/sparc/kmdb/kctl/kctl_isadep.c
@@ -71,8 +71,8 @@ kctl_ktrap_install(int tl, void (*handler)(void))
 	extern uint32_t kmdb_trap, kmdb_trap_tl1;
 	uint32_t *entryp = tl ? &kmdb_trap_tl1 : &kmdb_trap;
 	uint32_t *savp = tl ? &kctl_trap_brsav : &kctl_trap_tl1_brsav;
-	uint32_t hi = (uint32_t)handler >> 10;
-	uint32_t lo = (uint32_t)handler & 0x3ff;
+	uint32_t hi = (uint32_t)(uintptr_t)handler >> 10;
+	uint32_t lo = (uint32_t)(uintptr_t)handler & 0x3ff;
 	uint32_t inst;
 
 	*savp = *entryp;
